+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+* 1.1.0 ->
+  * added patch for new types:
+    * TeleportWorld (portal)
+    * PickableItem 
+    * Container (chests)
+  * added new icon "portal"
+  * updated vanilla configs with new pins
 * 1.0.4 -> replaced tree, dungeon and dot icon with less bold ones
 * 1.0.3 -> fixed missing tree icons
 * 1.0.1 & .2 ->